requirements.txt
flask
PyPDF2
openai
langchain_openai
faiss-cpu
llama_index
python-dotenv
Flask-Session